+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/catalystsquad/prometheus-salesforce-exporter/cmd"
+import "github.com/catalystcommunity/prometheus-salesforce-exporter/cmd"
 
 func main() {
 	cmd.Execute()