+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 
 ### Features
 
-* initial prometheus-salesforce-exporter ([#1](https://github.com/catalystsquad/prometheus-salesforce-exporter/issues/1)) ([274f86f](https://github.com/catalystsquad/prometheus-salesforce-exporter/commit/274f86fea1243c359a8b946a149f61b6a136ba15))
+* initial prometheus-salesforce-exporter ([#1](https://github.com/catalystcommunity/prometheus-salesforce-exporter/issues/1)) ([274f86f](https://github.com/catalystcommunity/prometheus-salesforce-exporter/commit/274f86fea1243c359a8b946a149f61b6a136ba15))